[Feat] 6차 배포

+++ src/main/java/outfoot/outfootserver/checkpage/dto/CheckPageCountListDto.java
@@ -6,11 +6,11 @@ import jakarta.validation.constraints.NotNull;
 import java.util.List;
 
 public record CheckPageCountListDto (
-        @Schema(description = "도장판 아이디", example = "1") @NotNull Long total,
+        @Schema(description = "도장판 아이디", example = "1") @NotNull int total,
         @Schema(description = "도장판 아이디", example = "1") List<CheckPageListResponse> checkPages
         ) {
 
-        public CheckPageCountListDto(@NotNull Long total, List<CheckPageListResponse> checkPages) {
+        public CheckPageCountListDto(@NotNull int total, List<CheckPageListResponse> checkPages) {
                 this.total = total;
                 this.checkPages = checkPages;
         }
